OPIK-118 rename FreeTierLimit* to UsageLimit*
