+++ apps/opik-backend/src/main/java/com/comet/opik/infrastructure/auth/CacheService.java
@@ -1,6 +1,6 @@
 package com.comet.opik.infrastructure.auth;
 
-import com.comet.opik.infrastructure.freetierlimit.Quota;
+import com.comet.opik.infrastructure.usagelimit.Quota;
 import lombok.Builder;
 
 import java.util.List;
