+++ apps/opik-backend/src/main/java/com/comet/opik/infrastructure/usagelimit/Quota.java
@@ -1,4 +1,4 @@
-package com.comet.opik.infrastructure.freetierlimit;
+package com.comet.opik.infrastructure.usagelimit;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
