+++ apps/opik-backend/src/main/java/com/comet/opik/infrastructure/usagelimit/UsageLimited.java
@@ -1,4 +1,4 @@
-package com.comet.opik.infrastructure.freetierlimit;
+package com.comet.opik.infrastructure.usagelimit;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface FreeTierLimited {
+public @interface UsageLimited {
 }
